In this Ice boat Game project we are Enhancing it's graphics in 3D format.Various Algorithms are used in this project.
OpenGL is a very important open source library for programmers who wish
to work with graphics. OpenGL can work with 2D and 3D images. OpenGL
helps to get a virtual view of real world objects and also helps in
visualization of objects. Icicles game is an Open-GL based "old-type" 2D
game. It consists of a player in an environment where ice cubes are
falling from sky. If the player collides with any of the ice cube then
game is over. The player score is the no. of ice cube which he/she can
dodge.

This game highlights various concepts such as
• Collision-Detection Algorithm
• KeyBoard Input Effect in all the Menu
• Double Buffer to achieve flicker-free animation. High FPS (60 FPS)
• All Basic OpenGL function has been used
• Settings and high score saved using files